+++ util-core/src/test/scala/com/twitter/io/ReaderTest.scala
@@ -6,6 +6,7 @@ import com.twitter.util.{Await, Awaitable, Future, Promise}
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 import java.util.concurrent.atomic.AtomicBoolean
 import org.mockito.Mockito._
+import org.scalacheck.Gen
 import org.scalatest.concurrent.{Eventually, IntegrationPatience}
 import org.scalatest.prop.GeneratorDrivenPropertyChecks
 import org.scalatest.{FunSuite, Matchers}
@@ -73,13 +74,33 @@ class ReaderTest
     }
   }
 
+  test("Reader.chunked") {
+    val stringAndChunk = for {
+      s <- Gen.alphaStr
+      i <- Gen.posNum[Int].suchThat(_ <= s.length)
+    } yield (s, i)
+
+    forAll(stringAndChunk) { case (s, i) =>
+      val r = Reader.chunked(Reader.fromBuf(Buf.Utf8(s)), i)
+
+      def readLoop(): Unit = await(r.read(Int.MaxValue)) match {
+        case Some(b) =>
+          assert(b.length <= i)
+          readLoop()
+        case None => ()
+      }
+
+      readLoop()
+    }
+  }
+
   test("Reader.concat") {
-    forAll { (ss: List[String]) =>
+    forAll { ss: List[String] =>
       val readers = ss.map { s =>
         BufReader(Buf.Utf8(s))
       }
       val buf = Reader.readAll(Reader.concat(AsyncStream.fromSeq(readers)))
-      await(buf) should equal(Buf.Utf8(ss.mkString))
+      assert(await(buf) == Buf.Utf8(ss.mkString))
     }
   }